Put Channel and ChannelOnce into separate modules
