+++ crates/relayer-framework/src/base/one_for_all/impls/runtime/channel.rs
@@ -2,9 +2,9 @@ use async_trait::async_trait;
 
 use crate::base::core::traits::sync::Async;
 use crate::base::one_for_all::traits::runtime::{OfaRuntime, OfaRuntimeWrapper};
-use crate::base::runtime::traits::channel::{
-    CanCreateChannels, CanCreateChannelsOnce, CanUseChannels, CanUseChannelsOnce,
-    HasChannelOnceTypes, HasChannelTypes,
+use crate::base::runtime::traits::channel::{CanCreateChannels, CanUseChannels, HasChannelTypes};
+use crate::base::runtime::traits::channel_once::{
+    CanCreateChannelsOnce, CanUseChannelsOnce, HasChannelOnceTypes,
 };
 use crate::std_prelude::*;
