+++ crates/relayer-framework/src/base/runtime/traits/channel.rs
@@ -14,35 +14,8 @@ pub trait HasChannelTypes: HasErrorType {
         T: Async;
 }
 
-pub trait HasChannelOnceTypes: HasErrorType {
-    type SenderOnce<T>: Async
-    where
-        T: Async;
-
-    type ReceiverOnce<T>: Async
-    where
-        T: Async;
-}
-
 pub trait CanCreateChannels: HasChannelTypes {
     fn new_channel<T>() -> (Self::Sender<T>, Self::Receiver<T>)
-    where
-        T: Async;
-}
-
-pub trait CanCreateChannelsOnce: HasChannelOnceTypes {
-    fn new_channel_once<T>() -> (Self::SenderOnce<T>, Self::ReceiverOnce<T>)
-    where
-        T: Async;
-}
-
-#[async_trait]
-pub trait CanUseChannelsOnce: HasChannelOnceTypes {
-    fn send_once<T>(sender: Self::SenderOnce<T>, value: T) -> Result<(), Self::Error>
-    where
-        T: Async;
-
-    async fn receive_once<T>(receiver: Self::ReceiverOnce<T>) -> Result<T, Self::Error>
     where
         T: Async;
 }
