+++ crates/relayer-framework/src/full/batch/impls/spawn.rs
@@ -3,7 +3,8 @@ use core::marker::PhantomData;
 use crate::base::chain::traits::types::HasIbcChainTypes;
 use crate::base::relay::traits::target::ChainTarget;
 use crate::base::relay::traits::types::HasRelayTypes;
-use crate::base::runtime::traits::channel::{CanUseChannels, CanUseChannelsOnce};
+use crate::base::runtime::traits::channel::CanUseChannels;
+use crate::base::runtime::traits::channel_once::CanUseChannelsOnce;
 use crate::base::runtime::traits::log::{HasLogger, LevelDebug};
 use crate::base::runtime::traits::runtime::HasRuntime;
 use crate::base::runtime::traits::sleep::CanSleep;
